+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 <!--
-  Copyright (c) 2023 Michael Federczuk
+  Copyright (c) 2024 Michael Federczuk
   SPDX-License-Identifier: CC-BY-SA-4.0
 -->
 
@@ -10,6 +10,13 @@
 All notable changes to this project will be documented in this file.
 The format is based on [**Keep a Changelog v1.0.0**](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [**Semantic Versioning v2.0.0**](https://semver.org/spec/v2.0.0.html).
+
+## Unreleased ##
+
+### Changed ###
+
+* The ktlint version `1.2.0` is rejected because it contains
+  [a bug](https://github.com/pinterest/ktlint/issues/2578) that breaks the behavior of the Git pre-commit hook script
 
 ## [v0.1.0-indev08] - 2023-10-28 ##